Skip tee() calls if node has exactly 1 output edge
If node splits data between >= 2 edges, we have to use tee() to split
the data between interested pipes, and we then splice the data to
/dev/null when we are sure it has been consumed by each pipe. However,
if the data is only going to 1 edge i.e. it does not need to be split,
then splice directly to that pipe. Avoids the overhead of duplicating
the data, splicing the original to /dev/null, ...