Stop spinning on failed writes
Previously, an event was fired whenever an out pipe from a node was
readable. If we were unable to write to an output, we would exit the
callback and the event would immediately fire again, leading to the
program spinning. Now, whenever the out pipe is readable, create events
for the in pipe which fire when writable. Only re-add the readable event
once all pipes have been subjected to the writable callback.

On an example long running task, reduced cpu usage from 95+% to ~15%.